Renamed TagSelector file and added 4 tags
