+++ src/views/SharedComponents/TagSelector.js
@@ -23,18 +23,22 @@ const TagSelector = ({handleSelect, handleDeselect, defaultValue, value, classNa
         <Option key="Academics"> Academics </Option>
         <Option key="Art"> Art </Option>
         <Option key="Athletics"> Athletics </Option>
+        <Option key="Awareness">Awareness</Option>
         <Option key="Business"> Business </Option>
         <Option key="Competition"> Competition </Option>
         <Option key="Cultural"> Cultural </Option>
         <Option key="Cooking"> Cooking </Option>
         <Option key="Diversity"> Diversity </Option>
         <Option key="Gaming"> Gaming </Option>
+        <Option key="Medicine">Medicine</Option>
         <Option key="Mentorship"> Mentorship </Option>
         <Option key="Politics"> Politics </Option>
         <Option key="Programming"> Programming </Option>
         <Option key="Recreation">Recreation </Option>
         <Option key="Religion">Religion</Option>
         <Option key="Science">Science</Option>
+        <Option key="Service">Service</Option>
+        <Option key="Sports">Sports</Option>
         <Option key="Theater"> Theater </Option>
         <Option key="Technology"> Technology </Option>
         <Option key="Volunteering"> Volunteering </Option>
